seperate processing logic into tokenservice
